fix: migrate table assigned tasks

+++ database/migrations/2023_05_09_131026_create_assigned_tasks_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('attached_file')->nullable();
             $table->integer('bonus')->default(0);
             $table->string('comment_moderator')->nullable();
-            $table->unsignedInteger('user_id_moderator')->nullable();
+            $table->unsignedBigInteger('user_id_moderator')->nullable();
             $table->foreign('user_id_moderator')->references('id')->on('users')->cascadeOnUpdate()->cascadeOnDelete();
             $table->enum('status', [array_column(AssignedTaskStatus::cases(), 'value')])->default(AssignedTaskStatus::Performed->value);
             $table->timestamps();
